+++ syntax/tests/printer/expr/block.js
@@ -16,10 +16,10 @@ let () = {
 let reifyStyle = (type a, x: 'a): (style<a>, a) => {
   module Internal = {
     type rec constructor
-    @bs.val external canvasGradient: constructor = "CanvasGradient" /* internal */
-    @bs.val external canvasPattern: constructor = "CanvasPattern" /* internal */
+    @val external canvasGradient: constructor = "CanvasGradient" /* internal */
+    @val external canvasPattern: constructor = "CanvasPattern" /* internal */
     let instanceOf = (
-      %bs.raw(`function(x,y) {return +(x instanceof y)}`): ('a, constructor) => bool
+      %raw(`function(x,y) {return +(x instanceof y)}`): ('a, constructor) => bool
     ) /* internal */
   }