+++ syntax/tests/printer/structure/extension.js
@@ -1,10 +1,10 @@
-%%bs.raw("__eval__gc()")
+%%raw("__eval__gc()")
 
 @attr
-%%bs.raw("__eval__gc()")
+%%raw("__eval__gc()")
 
 @attrStructureLvl @attrStructureLvl2 @attrStructureLvl3
-%%bs.raw("__eval__gc()")
+%%raw("__eval__gc()")
 
 @attrStructureLvl @attrStructureLvl2 @attrStructureLvl3 @attrStructureLvl4 @attrStructureLvl5
-%%bs.raw("__eval__gc()")
+%%raw("__eval__gc()")